feat(backend): admin orders cleanup — W19 force-cancel + dead_letter + TD-MONEY-01
W19 (force-cancel side-effects in admin/orders.py):
- when force-status target is a cancel state and order had successful
  pay, issue full auto-refund; on failure record dead_letter row so
  ops can compensate without losing the original transition
- notify both parties (patient always; companion when assigned) when
  forcing a cancel state; failures fall through to dead_letter
- when force-completing, bump companion CompanionProfile.total_orders
  so manual completions stay in sync with the review path counter
- audit reason now appends side_effects=... summary

dead_letter (W19 ops queue):
- new app/models/dead_letter.py with status enum + JSONB payload
- alembic c1d2e3f4a5b7 (PG: explicit DO$$ CREATE TYPE pattern; SQLite
  falls back to String/JSON so tests still bring schema up)
- new app/services/dead_letter.py best-effort writer (never raises)
- new admin endpoints under /api/v1/admin/dead-letters:
  list (filter status/channel/target_id, paginated),
  detail, resolve (stamps operator + writes admin_audit_logs row)
- cancel.py::reject_order auto-refund failure path now writes a
  dead_letter via a fresh session (the primary tx is rolled back by
  the BadRequestException propagation)

TD-MONEY-01 (early; deadline 2026-06-30):
- removed float() coercion from schemas/order.py price + amount and
  schemas/companion.py total_earnings; Pydantic v2 default emits
  Decimal as string in JSON, preserving precision (ADR-0030)
- updated unit/test_decimal_money.py contract assertion to expect
  string emit; updated affected JSON-response assertions in
  test_orders / test_refund_e2e / test_companion_stats /
  test_full_lifecycle to compare against the canonical "NNN.NN" form

Client compat:
- wechat utils/formatCurrency.js uses Number(value) — accepts both
- admin-h5 fmtAmount() likewise
- iOS Order.price / Payment.amount are Swift Decimal which accepts
  numeric tokens; CompanionStatsResponse.totalEarnings is Double and
  may need a follow-up to accept string — see iOS issue note in PR

+++ backend/app/schemas/order.py
@@ -110,11 +110,13 @@ class OrderResponse(BaseModel):
         return out
 
     @field_serializer("price")
-    def _ser_price(self, v: Decimal) -> float:
-        # ADR-0030: 内部 Decimal，对外保持 number（前端契约不破）
-        # TODO(W19, deadline 2026-06-30 / W26): remove float() coercion once mobile
-        # clients ship Decimal-aware parsers. Tracking: TD-MONEY-01.
-        return float(Decimal(v).quantize(Decimal("0.01")))
+    def _ser_price(self, v: Decimal) -> Decimal:
+        # ADR-0030 / TD-MONEY-01 (done 2026-05-25): 不再转 float。
+        # Pydantic v2 默认将 Decimal 以字符串输出为 JSON，避免
+        # 浮点精度丢失。客户端（wechat / admin-h5 / ios）均能
+        # 处理 "199.00" 形式；wechat / admin-h5 走 Number()，ios
+        # 使用 DecimalString 解码包装。
+        return Decimal(v).quantize(Decimal("0.01"))
 
 
 class OrderListResponse(BaseModel):
@@ -134,8 +136,6 @@ class PaymentResponse(BaseModel):
     model_config = {"from_attributes": True}
 
     @field_serializer("amount")
-    def _ser_amount(self, v: Decimal) -> float:
-        # ADR-0030: 内部 Decimal，对外保持 number
-        # TODO(W19, deadline 2026-06-30 / W26): remove float() coercion once mobile
-        # clients ship Decimal-aware parsers. Tracking: TD-MONEY-01.
-        return float(Decimal(v).quantize(Decimal("0.01")))
+    def _ser_amount(self, v: Decimal) -> Decimal:
+        # ADR-0030 / TD-MONEY-01 (done 2026-05-25): 不再转 float。
+        return Decimal(v).quantize(Decimal("0.01"))
